move all step schema types do types folder
