+++ packages/types/src/documents/app/automation/index.ts
@@ -1,0 +1,2 @@
+export * from "./automation"
+export * from "./schema"
